command_line/requirements.txt
click==8.0.4
cryptography==41.0.3
gradio==4.27.0
matplotlib==3.7.2
numpy==1.24.3
Pillow==9.4.0
